WEBAPP-142: Move Error to components
