+++ src/modules/common/components/__tests__/Error.spec.js
